requirements.txt
ldap3>=2.9.1
python-ldap>=3.4.3
coldfront
